Clean up landing app for Casimir V1
