+++ apps/landing/src/App.vue
@@ -1,12 +1,6 @@
-<script lang="ts" setup>
-import { RouterView } from 'vue-router'
-import DefaultLayout from '@/layouts/default-layout.vue'
-</script>
-
 <template>
-  <Suspense>
-    <component :is="DefaultLayout">
-      <RouterView />
-    </component>
-  </Suspense>
+  <div>
+    <!-- No Layout needed -->
+    <RouterView />
+  </div>
 </template>
