+++ apps/landing/src/index.css
@@ -1,3 +1,32 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+:root {
+
+    font-synthesis: none;
+    text-rendering: optimizeLegibility;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    -webkit-text-size-adjust: 100%;
+  }
+  
+  #app {
+    max-width: 1280px;
+    margin: 0 auto;
+    height: 100vh;
+    background: #fff;
+  }
+
+
+@layer base {
+    /* Base Tags */
+}
+
+@layer components {
+    /* Component Classes */
+}
+
+@layer utilities {
+    /* Utilities Styling */
+}
