--- apps/landing/src/layouts/default-layout.vue
@@ -1,17 +1,0 @@
-<template>
-  <div
-    id="layout"
-  >
-    <!-- Nav -->
-    <!-- <NavBar /> -->
-
-    <!-- Page -->
-    <div id="content">
-      <slot />
-    </div>
-  </div>
-</template>
-
-<script lang="ts" setup>
-// import NavBar from '@/components/nav-bar.vue'
-</script>
