+++ apps/landing/src/pages/community/Community.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    Community
+  </div>
+</template>
