+++ apps/landing/src/pages/home/Home.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    Home
+  </div>
+</template>
