+++ apps/landing/src/pages/not-found/NotFound.vue
@@ -1,0 +1,6 @@
+<template>
+  <div>
+    Not Found
+  </div>
+</template>
+
