+++ apps/landing/src/pages/whitepaper/Whitepaper.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    Whitepaper
+  </div>
+</template>
