+++ apps/landing/tailwind.config.js
@@ -3,14 +3,6 @@ module.exports = {
         './**/*.{vue,ts,js,css}'
     ],
     theme: {
-        extend: {
-            colors: {},
-            screens: {
-                'laptop': '450px',
-                'tablet': '450px',
-                'mobile': '450px'
-            },
-        }
     },
     plugins: []
 }
