+++ apps/landing/tsconfig.json
@@ -21,5 +21,5 @@
       "./node_modules/@types"
     ]
   },
-  "include": ["./src/**/*.ts", "./src/**/*.vue"]
+  "include": ["./src/**/*.ts", "./src/**/*.vue", "src/main.js", "src/composables/router.js", "src/composables/router.js", "src/main.js"]
 }
