+++ package.json
@@ -14,7 +14,7 @@
     "deploy": "scripts/cdk/deploy -d infrastructure/cdk",
     "dev": "scripts/local/dev -f \"$npm_config_fork\" -m \"$npm_config_mock\" -n \"$npm_config_network\" -s \"$npm_config_speculos\" -t \"$npm_config_tunnel\"",
     "dev:ethereum": "scripts/ethereum/dev -f \"$npm_config_fork\"",
-    "dev:landing": "scripts/local/dev -a \"landing\" -m \"$npm_config_mock\"",
+    "dev:landing": "npm run dev --workspace @casimir/landing",
     "lint": "eslint --ext .vue,.ts ./ --fix",
     "test": "echo \"Error: no test specified\" && exit 1",
     "test:ethereum": "scripts/ethereum/test -f \"$npm_config_fork\"",
